+++ packages/core-js/es/number/to-exponential.js
@@ -1,0 +1,4 @@
+require('../../modules/es.number.to-exponential');
+var entryUnbind = require('../../internals/entry-unbind');
+
+module.exports = entryUnbind('Number', 'toExponential');