+++ packages/core-js/es/number/virtual/index.js
@@ -1,3 +1,4 @@
+require('../../../modules/es.number.to-exponential');
 require('../../../modules/es.number.to-fixed');
 require('../../../modules/es.number.to-precision');
 var entryVirtual = require('../../../internals/entry-virtual');