+++ packages/core-js/modules/es.math.log10.js
@@ -1,12 +1,8 @@
 var $ = require('../internals/export');
-
-var log = Math.log;
-var LOG10E = Math.LOG10E;
+var log10 = require('../internals/math-log10');
 
 // `Math.log10` method
 // https://tc39.es/ecma262/#sec-math.log10
 $({ target: 'Math', stat: true }, {
-  log10: function log10(x) {
-    return log(x) * LOG10E;
-  }
+  log10: log10
 });