+++ packages/core-js/stable/number/virtual/to-exponential.js
@@ -1,0 +1,3 @@
+var parent = require('../../../es/number/virtual/to-exponential');
+
+module.exports = parent;