+++ tests/tests/es.number.to-fixed.js
@@ -59,9 +59,9 @@ QUnit.test('Number#toFixed', assert => {
   assert.same(new Number(1e21).toFixed(NaN), String(1e21));
   assert.same(new Number(1e21).toFixed('some string'), String(1e21));
   assert.notThrows(() => new Number(1e21).toFixed(-0.1) === String(1e21));
-  assert.throws(() => 1.0.toFixed(-101), RangeError, 'If f < 0 or f > 20, throw a RangeError exception.');
-  assert.throws(() => 1.0.toFixed(101), RangeError, 'If f < 0 or f > 20, throw a RangeError exception.');
-  assert.throws(() => NaN.toFixed(Infinity), RangeError, 'If f < 0 or f > 20, throw a RangeError exception.');
+  assert.throws(() => 1.0.toFixed(-101), RangeError, 'If f < 0 or f > 20 (100), throw a RangeError exception.');
+  assert.throws(() => 1.0.toFixed(101), RangeError, 'If f < 0 or f > 20 (100), throw a RangeError exception.');
+  assert.throws(() => NaN.toFixed(Infinity), RangeError, 'If f < 0 or f > 20 (100), throw a RangeError exception.');
   assert.throws(() => toFixed.call({}, 1), TypeError, '? thisNumberValue(this value)');
   assert.throws(() => toFixed.call('123', 1), TypeError, '? thisNumberValue(this value)');
   assert.throws(() => toFixed.call(false, 1), TypeError, '? thisNumberValue(this value)');